requirements.txt
certifi>=2025.1.31
charset-normalizer>=3.4.1
decorator>=5.2.1
filelock>=3.18.0
fsspec>=2025.3.0
idna>=3.10
imageio>=2.37.0
imageio-ffmpeg>=0.6.0
Jinja2>=3.1.6
llvmlite>=0.44.0
MarkupSafe>=3.0.2
more-itertools>=10.6.0
moviepy>=2.1.2
mpmath>=1.3.0
networkx>=3.4.2
numba>=0.61.0
numpy<2.0.0,>=1.21.2
openai-whisper>=20240930
pillow>=10.4.0
proglog>=0.1.10
pydub>=0.25.1
python-dotenv>=1.0.1
regex>=2024.11.6
requests>=2.32.3
sympy>=1.13.3
tiktoken>=0.9.0
torch>=2.2.2
tqdm>=4.67.1
typing_extensions>=4.12.2
urllib3>=2.3.0
